Namespacing of the API caching and small refactor of the Store classes.
This will simplify the eventual extraction to a separated gem.
