+++ lib/caching/caching.rb
@@ -1,0 +1,9 @@
+module Caching
+  class << self
+    cattr_accessor :backend
+
+    def method_missing(method, *args, &block)
+      backend.current.send(method, *args, &block)
+    end
+  end
+end
